remove confusing "manually merge layers together"
This sentence confused me when I was reading it for the first time. What
does it mean to "manually merge layers together"?

Maybe it's intended to be a header for the following paragraphs, but
it's not styled as such and actually, it's explained pretty well farther
down the document, so I think it's all together better with it removed.